Move TopoFetcher to argopy.related
