+++ argopy/__init__.py
@@ -35,12 +35,13 @@ from . import errors  # noqa: E402
 from . import plot  # noqa: E402
 from .plot import dashboard, ArgoColors  # noqa: E402
 from .utilities import show_versions, show_options, clear_cache, lscache  # noqa: E402
-from .utilities import TopoFetcher, ArgoNVSReferenceTables, OceanOPSDeployments, ArgoDocs  # noqa: E402
+from .utilities import ArgoNVSReferenceTables, OceanOPSDeployments, ArgoDocs  # noqa: E402
 from .utilities import monitor_status as status  # noqa: E402
 from .options import set_options, reset_options  # noqa: E402
 from .data_fetchers import CTDRefDataFetcher  # noqa: E402
 from .stores import ArgoIndex  # noqa: E402
 from .utils import compute  # noqa: E402, F401
+from .related import TopoFetcher # noqa: E402
 
 #
 __all__ = (
