+++ argopy/related/__init__.py
@@ -1,0 +1,7 @@
+from .topo import TopoFetcher
+
+#
+__all__ = (
+    # Classes:
+    "TopoFetcher",
+)
